RFC: Add the ability to make an existing struct an overlay pass (#19)
The macro adds a bunch of methods now, so if we want to use our
own struct (e.g. to add some fields), need need to be able to
pass in the name of the struct. An alternative design would
be to put the macro on the struct definition or to make the
methods all on the abstract type, so that any subtype is automatically
enrolled.