fix world age errors for `AbstractBindingOverlay` on v1.12
Starting from v1.12.1, the implementation of `isconst/getglobal` became
more accurate, and when called in a generator's world that was defined
before the target method table, it now correctly determines that the
method table is not yet defined.

So, from v1.12 onwards, it seems necessary to use functions like
`[getglobal|isconst]_at_world` as implemented here. This might be
something that should be provided by Base itself though.